refactor: isolate WASM component handling into dedicated module
